fixed the naming conventions
My bad!
